xen/blktap: Fix device destruction path.
Blockdev teardown used to be piggybacked on VMA shutdown. This implied
we're failing requests while bdevs are still open.

Let userspace try to remove it, yielding -EAGAIN until bd_openers
drops to zero. Unblock ring .poll whenever we're ready to actually
close.

This cleans up the backward compat sysfs remove. We destroy the sysfs
device synchronously, unblocking the removing thread to avoid
deadlock. This is normally done by tapdisk. If the ring is not active,
schedule a workqueue item instead. Which is awkward, but guarantees the
minor is fully released by the time the remover returns.

Signed-off-by: Daniel Stodden <daniel.stodden@citrix.com>
Signed-off-by: Jake Wires <jake.wires@citrix.com>